# assets/openapi.yaml
openapi: 3.1.0
info:
  title: openccu-loom REST API
  version: 1.0.0
  description: |
    North-bound REST + WebSocket surface of the openccu-loom daemon.
    Errors follow RFC 9457 problem+json.
  license:
    name: MIT
    url: https://opensource.org/license/mit/
servers:
  - url: /api/v1
security:
  - basicAuth: []
  - bearerAuth: []
  - openIdConnect: [viewer]
...
paths:
...
  /devices/{addr}/cdps:
    get:
      operationId: listCustomDataPoints
      summary: List all custom data points of a device
      tags: [devices]
      parameters:
        - { in: path, name: addr, required: true, schema: { type: string } }
      responses:
        "200":
          description: Custom DP list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CustomDPSummary" }
        "404": { $ref: "#/components/responses/NotFound" }
  /devices/{addr}/cdps/{name}:
    get:
      operationId: getCustomDataPoint
      summary: Get a single custom data point by name
      tags: [devices]
      parameters:
        - { in: path, name: addr, required: true, schema: { type: string } }
        - { in: path, name: name, required: true, schema: { type: string } }
      responses:
        "200": { description: Custom DP detail }
        "404": { $ref: "#/components/responses/NotFound" }
  /devices/{addr}/cdps/{name}/{operation}:
    post:
      operationId: invokeCustomDataPoint
      summary: "Invoke an operation on a custom data point (turn_on, set_temperature, open, …)"
      tags: [devices]
      security:
        - basicAuth: []
        - bearerAuth: []
        - openIdConnect: [operator]
      parameters:
        - { in: path, name: addr, required: true, schema: { type: string } }
        - { in: path, name: name, required: true, schema: { type: string } }
        - in: path
          name: operation
          required: true
          schema:
            type: string
            description: "Operation name, e.g. turn_on, turn_off, set_temperature, open, close, …"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomDPInvokeRequest" }
      responses:
        "204": { description: Accepted }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/BadRequest" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
  /audit:
    get:
      operationId: listAudit
      summary: Recent change history (FIFO buffer)
      tags: [audit]
      parameters:
        - in: query
          name: device
          schema: { type: string }
          description: >
            Filter by device address prefix (case-insensitive).
            E.g. ?device=0001 returns entries whose DeviceAddress starts with "0001".
        - in: query
          name: op
          schema: { type: string }
          description: >
            Filter by action/source tag substring (case-insensitive).
            E.g. ?op=paramset matches paramset_write, link_paramset_write, etc.
        - in: query
          name: since
          schema: { type: string, format: date-time }
          description: >
            Only return entries with timestamp >= this value (inclusive, RFC3339).
            Returns 400 if the value cannot be parsed.
        - in: query
          name: until
          schema: { type: string, format: date-time }
          description: >
            Only return entries with timestamp < this value (exclusive, RFC3339).
            Returns 400 if the value cannot be parsed.
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 10000, default: 1000 }
          description: >
            Maximum entries to return. 0 or absent uses the default (1000).
            Values above 10000 are capped to 10000.
      responses:
        "200":
          description: Filtered list of audit entries (newest first).
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AuditEntry" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: |
        Username + password (operator-grade or above for mutations).
        The role hierarchy (viewer ⊂ operator ⊂ admin) is enforced
        server-side regardless of authentication scheme.
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        API token issued via the auth tokens surface (admin-only
        provisioning). Token's stored role determines authorisation.
    openIdConnect:
      type: openIdConnect
      description: |
        OpenID Connect authentication. The operator configures the
        OIDC discovery URL; the placeholder below is the standard
        well-known path. Per-operation `security:` blocks express
        the **minimum role** the daemon enforces via OIDC-style
        scopes — the same enforcement applies when the client
        authenticates via basicAuth or bearerAuth, but those schemes
        cannot natively express scopes in OpenAPI 3.x.

        Scope vocabulary mirrors `Identity.role`:
          - `viewer` — read-only operations
          - `operator` — mutations (write data points, manage links,
            edit schedules, run config-import, trigger refresh)
          - `admin` — dangerous operations (delete device, edit log
            levels, list users/tokens, trigger backup, manage
            interfaces, edit Matter fabrics)
      openIdConnectUrl: /.well-known/openid-configuration
...
  responses:
    BadRequest:
      description: Validation or parsing failure
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    BadGateway:
      description: Upstream error (CCU unavailable, write failed)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    ServiceUnavailable:
      description: Backend not wired
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    CustomDPInvokeRequest:
      type: object
      description: "Body for POST /cdps/{name}/{operation}. The operation is in the URL path."
      properties:
        params:
          type: object
          additionalProperties: true
          description: "Operation-specific parameters (category-dependent)."
        priority:
          type: string
          enum: [critical, high, default, low]
    CustomDPSummary:
      type: object
      required: [name, category, channel_no, supported_operations]
      properties:
        name: { type: string }
        category: { type: string }
        channel_no: { type: integer }
        supported_operations:
          type: array
          items: { type: string }
        kind:
          type: string
          description: |
            Stable widget hint derived from the concrete Custom-DP
            type — e.g. `light`, `light_color`, `light_color_temp`,
            `light_fixed_color`, `light_rgbw`, `cover_blind`,
            `cover_garage`, `climate_hmip`, `climate_rf`,
            `climate_simple`, `lock`, `siren`, `switch`,
            `text_display`. Drives the SPA's CDP-aware widget
            selection (ADR 0016). Empty when the kind classifier
            does not recognise the Custom-DP type.
        channels:
          type: array
          items: { type: integer }
          description: |
            CCU channels this Custom-DP is materialised on. The
            current implementation always contains exactly the
            primary channel (matching `channel_no`); future
            iterations may expand to include sibling channels
            pulled in via the `RebasedChannelGroupConfig`
            mechanism.
        capabilities:
          type: object
          additionalProperties: { type: boolean }
          description: |
            Flat string→bool map of optional features the device
            supports — e.g. `dimmable`, `color`, `color_temp`,
            `effects`, `tilt`, `boost`, `away`. Mirrors the
            per-category Capability struct from
            `internal/model/custom/mixins.go`.
    Problem:
      type: object
      description: |
        RFC 9457 problem detail. The `type` URI selects one of the
        canonical problem kinds the daemon emits; the `code` member
        carries the short tag (also surfaced as the `X-Problem-Code`
        response header). Clients should switch on `code` for typed
        error handling — the `type` URI is the human-readable form.

        **Rate-limiting:** when the daemon emits `code: rate_limited`
        (HTTP 429) it sets the `Retry-After` response header per
        RFC 9110 §10.2.3 with the integer number of seconds the
        client should wait before retrying. Clients that respect
        the header are immune to the throttle; clients that don't
        will see repeated 429s with the same `Retry-After` value
        until the throttle window expires.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: |
            Stable URI for one of the canonical problem kinds. All
            URIs share the base `https://openccu-loom.dev/errors/`
            and the trailing segment matches the `code` enum below.
            The base URI is a stable identifier, not a fetchable
            resource — it is not guaranteed to resolve to a document.
          enum:
            - https://openccu-loom.dev/errors/validation
            - https://openccu-loom.dev/errors/not_found
            - https://openccu-loom.dev/errors/conflict
            - https://openccu-loom.dev/errors/unauthorized
            - https://openccu-loom.dev/errors/forbidden
            - https://openccu-loom.dev/errors/unsupported
            - https://openccu-loom.dev/errors/rate_limited
            - https://openccu-loom.dev/errors/internal
            - https://openccu-loom.dev/errors/bad_request
            - https://openccu-loom.dev/errors/service_unready
            - https://openccu-loom.dev/errors/upstream_unavailable
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: Short tag (also surfaced as the `X-Problem-Code` response header).
          enum:
            - validation
            - not_found
            - conflict
            - unauthorized
            - forbidden
            - unsupported
            - rate_limited
            - internal
            - bad_request
            - service_unready
            - upstream_unavailable
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              reason: { type: string }
              min: {}
              max: {}
...
    AuditEntry:
      type: object
      required: [timestamp, action]
      description: One recorded user-initiated configuration change.
      properties:
        timestamp:
          type: string
          format: date-time
          description: UTC time at which the change was recorded.
        user:
          type: string
          description: Authenticated user who triggered the change (empty if unauthenticated).
        action:
          type: string
          description: >
            Stable action tag, e.g. paramset_write, link_add, data_point_write.
            New values may be added in future releases.
        device_address:
          type: string
          description: Primary device or channel address, if applicable.
        channel_no:
          type: integer
          description: Channel number within the device, if applicable.
        paramset:
          type: string
          description: Paramset type (MASTER, VALUES, LINK), if applicable.
        peer:
          type: string
          description: Peer address for link operations.
        parameter:
          type: string
          description: Parameter name for single-parameter writes.
        note:
          type: string
          description: Free-form context (e.g. profile ID for active_profile changes).
        changes:
          type: array
          description: Per-parameter before/after pairs for paramset writes.
          items:
            type: object
            required: [parameter]
            properties:
              parameter: { type: string }
              before: {}
              after: {}
...